backend/requirements.txt
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
firebase-admin
google-cloud-firestore
fuzzywuzzy
python-Levenshtein
numpy
pytz
sendgrid